pyproject.toml
[project]
name = "vv-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "mysql-connector-python>=9.5.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "pymediainfo>=7.0.1",
    "scikit-learn>=1.7.2",
    "torch>=2.9.0",
    "uvicorn>=0.38.0",
    "chromadb>=0.5.0",
    "open-clip-torch>=2.24.0",
]